Fix for Bug 92630: Null Pointer Exception thrown by ajc compiler

+++ weaver/src/org/aspectj/weaver/bcel/UnwovenClassFile.java
@@ -167,8 +167,10 @@ public class UnwovenClassFile {
 				public boolean accept(File dir, String name) {
 					return name.startsWith(targetPrefix);
 				}});
-			for (int i = 0; i < weaverGenerated.length; i++) {
-				weaverGenerated[i].delete();
+			if (weaverGenerated!=null) {
+				for (int i = 0; i < weaverGenerated.length; i++) {
+					weaverGenerated[i].delete();
+				}
 			}
 		}
 		victim.delete();
